Remove outdated and unmaintained content
This commit removes pages which we are not maintaining, and probably should not even attempt to maintain.

Bounty list: we are not getting PRs for  listings
Notifications: there are far too many sources to list
Security EIPs: this is very subjective, and listing them would require
too much effort, duplicating conversations that happen elsewhere.